requirements.txt
python-dotenv
openai
openai-whisper
streamlit
pytube